Update font-firacode to v5.2
Summary:
Removes the OTF dist in favor of the TTF dist, as the OTF dist does not support hinting.

Notable changes in v5:
- Added several unicode characters, including 2-em dash and 3-em dash
- Fixed arrows next to lowercase letters
- Vertically-aligned colons and greater/less than signs
- Disable =/ due to conflicts with paths
- Added a couple new ligatures

v5.1 fixes an issue with the 3-em dash, and v5.2 fixes an issue with the TTF build.

Full changelog for v5 is available [here](https://github.com/tonsky/FiraCode/releases/tag/5).

Test Plan: Installed and verified that ligatures work in libreoffice

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9029